+++ src/main/java/kitchenpos/menu_group/application/MenuGroupService.java
@@ -1,11 +1,15 @@
 package kitchenpos.menu_group.application;
 
 import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.dto.MenuGroupRequest;
+import kitchenpos.menu_group.dto.MenuGroupResponse;
 import kitchenpos.menu_group.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 public class MenuGroupService {
@@ -16,12 +20,15 @@ public class MenuGroupService {
     }
 
     @Transactional
-    public MenuGroup create(final MenuGroup menuGroup) {
-        return menuGroupRepository.save(menuGroup);
+    public MenuGroupResponse create(final MenuGroupRequest menuGroup) {
+        MenuGroup savedMenuGroup = menuGroupRepository.save(menuGroup.toEntity());
+        return new MenuGroupResponse(savedMenuGroup);
     }
 
     @Transactional(readOnly = true)
-    public List<MenuGroup> list() {
-        return menuGroupRepository.findAll();
+    public List<MenuGroupResponse> list() {
+        return menuGroupRepository.findAll().stream()
+                .map(MenuGroupResponse::new)
+                .collect(toList());
     }
 }
