+++ src/main/java/kitchenpos/menu_group/dto/MenuGroupRequest.java
@@ -1,0 +1,26 @@
+package kitchenpos.menu_group.dto;
+
+import kitchenpos.menu_group.domain.MenuGroup;
+
+public class MenuGroupRequest {
+    private String name;
+
+    public MenuGroupRequest() {
+    }
+
+    public MenuGroupRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public MenuGroup toEntity() {
+        return new MenuGroup(name);
+    }
+}
