+++ src/main/java/kitchenpos/menu_group/dto/MenuGroupResponse.java
@@ -1,0 +1,29 @@
+package kitchenpos.menu_group.dto;
+
+import kitchenpos.menu_group.domain.MenuGroup;
+
+public class MenuGroupResponse {
+    private Long id;
+    private String name;
+
+    public MenuGroupResponse(MenuGroup created) {
+        this.id = created.getId();
+        this.name = created.getName();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
