+++ src/main/java/kitchenpos/menu_group/ui/MenuGroupRestController.java
@@ -1,7 +1,8 @@
 package kitchenpos.menu_group.ui;
 
 import kitchenpos.menu_group.application.MenuGroupService;
-import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.dto.MenuGroupRequest;
+import kitchenpos.menu_group.dto.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,59 +22,17 @@ public class MenuGroupRestController {
 
     @PostMapping("/api/menu-groups")
     public ResponseEntity<MenuGroupResponse> create(@RequestBody final MenuGroupRequest request) {
-        final MenuGroup created = menuGroupService.create(request.toEntity());
+        final MenuGroupResponse created = menuGroupService.create(request);
         final URI uri = URI.create("/api/menu-groups/" + created.getId());
         return ResponseEntity.created(uri)
-                .body(new MenuGroupResponse(created))
+                .body(created)
                 ;
     }
 
     @GetMapping("/api/menu-groups")
-    public ResponseEntity<List<MenuGroup>> list() {
+    public ResponseEntity<List<MenuGroupResponse>> list() {
         return ResponseEntity.ok()
                 .body(menuGroupService.list())
                 ;
-    }
-
-    static class MenuGroupRequest {
-        private String name;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public MenuGroup toEntity() {
-            return new MenuGroup(name);
-        }
-    }
-
-    static class MenuGroupResponse {
-        private Long id;
-        private String name;
-
-        public MenuGroupResponse(MenuGroup created) {
-            this.id = created.getId();
-            this.name = created.getName();
-        }
-
-        public Long getId() {
-            return id;
-        }
-
-        public void setId(Long id) {
-            this.id = id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
     }
 }
